added registerDto, service and modify controller
